feat(hana)!: HANA integration tests
HANA TINYINT is unsigned by default.

+++ packages/hana/src/_internal/data-types-overrides.ts
@@ -69,10 +69,10 @@ export class NOW extends BaseTypes.NOW {
 export class TINYINT extends BaseTypes.TINYINT {
   toSql(): string {
     if (this.options.unsigned) {
-      return 'SMALLINT';
+      return 'TINYINT';
     }
 
-    return 'TINYINT';
+    return 'SMALLINT';
   }
 }
